Separate POSFloatingControlView file
